refact:유저 스켸쥴링 위치 변경
