+++ src/main/java/swyp/qampus/login/service/impl/UserServiceImpl.java
@@ -1,8 +1,11 @@
 package swyp.qampus.login.service.impl;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import swyp.qampus.exception.CommonErrorCode;
 import swyp.qampus.exception.RestApiException;
 import swyp.qampus.login.entity.User;
@@ -19,6 +22,7 @@ import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
     private final QuestionRepository questionRepository;
@@ -38,5 +42,21 @@ public class UserServiceImpl implements UserService {
         return questions.stream()
                 .map(MyQuestionResponseDto::of)
                 .collect(Collectors.toList());
+    }
+
+    @Override
+    @Transactional
+    @Scheduled(cron = "1 0 0 1 * * ")
+    public void userResetMonthly() {
+        userRepository.resetMonthlyChoiceCnt();
+        log.info("유저 monthlyChoice 초기화");
+    }
+
+    @Override
+    @Transactional
+    @Scheduled(cron = "59 59 23 * * 7")
+    public void userResetWeekly() {
+        userRepository.resetWeeklyChoiceCnt();
+        log.info("유저 weeklyChoice 초기화");
     }
 }
